+++ scripts/patch_recursive_apply.py
@@ -47,11 +47,14 @@ import subprocess
 
 # configuration options
 
-parser = argparse.ArgumentParser(description=__doc__)
+parser = argparse.ArgumentParser(
+    description=__doc__,
+    epilog=
+    "All additional arguments are passed to the patch command CMD.")
 parser.add_argument(
     'patch_cmd', metavar="CMD", type=str,
     help="use CMD to apply a patch between files; this command must "
-    "accept precisely 3 arguments: 'CMD oldfile newfile patchfile'")
+    "accept at least 3 arguments: 'CMD oldfile newfile patchfile ...'")
 parser.add_argument(
     'in_folder', type=str,
     help="folder containing original files")
@@ -61,7 +64,7 @@ parser.add_argument(
 parser.add_argument(
     'patch_folder', type=str,
     help="folder containing patch files")
-args = parser.parse_args()
+args, unknown_args = parser.parse_known_args()
 
 # actual script
 
@@ -75,7 +78,7 @@ def patch_cmd(in_file, out_file, patch_file):
         os.makedirs(folder)
     # apply patch_file if it exists
     if os.path.exists(patch_file):
-        command = [args.patch_cmd, in_file, out_file, patch_file]
+        command = [args.patch_cmd, in_file, out_file, patch_file] + unknown_args
         print("applying %s" % patch_file)
         subprocess.call(command)